add more data to the idempotent test
